+++ sqlmesh_dag_generator/__init__.py
@@ -2,7 +2,7 @@
 SQLMesh DAG Generator - Open Source Airflow Integration for SQLMesh
 """
 
-__version__ = "0.9.9"
+__version__ = "0.9.10"
 
 
 from sqlmesh_dag_generator.generator import SQLMeshDAGGenerator
@@ -12,6 +12,16 @@ from sqlmesh_dag_generator.airflow_utils import (
     register_credential_resolver,
     CredentialResolver,
 )
+from sqlmesh_dag_generator.airflow_compat import (
+    BaseHook,
+    BashOperator,
+    EmptyOperator,
+    PythonOperator,
+    TriggerDagRunOperator,
+    Variable,
+    dag_schedule_kwargs,
+    is_airflow_3,
+)
 
 __all__ = [
     "SQLMeshDAGGenerator",
@@ -20,6 +30,15 @@ __all__ = [
     "resolve_credentials",
     "register_credential_resolver",
     "CredentialResolver",
+    # AF2/AF3 shims (import from here or sqlmesh_dag_generator.airflow_compat)
+    "BaseHook",
+    "BashOperator",
+    "EmptyOperator",
+    "PythonOperator",
+    "TriggerDagRunOperator",
+    "Variable",
+    "dag_schedule_kwargs",
+    "is_airflow_3",
 ]